requirements.txt
annotated-types==0.7.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
dominate==2.9.1
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.5
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
platformdirs==4.3.8
protobuf==6.31.1
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.4
scipy==1.16.1
sentry-sdk==2.34.1
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sympy==1.14.0
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tornado==6.5.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
visdom==0.2.4
wandb==0.21.1
websocket-client==1.8.0
